+++ jiaozivideoplayer/src/main/java/cn/jzvd/JZMediaManager.java
@@ -64,15 +64,15 @@ public class JZMediaManager implements TextureView.SurfaceTextureListener {
         instance().jzMediaInterface.currentDataSource = currentDataSource;
     }
 
-    public static int getCurrentPosition() {
+    public static long getCurrentPosition() {
         return instance().jzMediaInterface.getCurrentPosition();
     }
 
-    public static int getDuration() {
+    public static long getDuration() {
         return instance().jzMediaInterface.getDuration();
     }
 
-    public static void seekTo(int time) {
+    public static void seekTo(long time) {
         instance().jzMediaInterface.seekTo(time);
     }
